Implement AirQuality board skets with new NodeManager version, Radically strip down the SensorMQ class
Remove everything that is not absolutely required for running. Switch to static const for values that in general might be nice to have configurable, but which will be the same for all our uses. This is to bring down the overall SRAM usage with 10 air quality sensors. It is NOT intended for the broad public, where configurability is predominant.